Folder rename resolved to service for backend
