+++ backend/src/main/java/com/oopproject/form/controllers/UserController.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import com.oopproject.form.models.User;
-import com.oopproject.form.services.UserService;
+import com.oopproject.form.service.UserService;
 
 @RestController
 @CrossOrigin
