+++ backend/src/main/java/com/oopproject/form/security/jwt/AuthTokenFilter.java
@@ -12,7 +12,7 @@ import org.springframework.security.web.authentication.WebAuthenticationDetailsS
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
 
-import com.oopproject.form.services.UserDetailsServiceImpl;
+import com.oopproject.form.service.UserDetailsServiceImpl;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
