+++ backend/src/main/java/com/oopproject/form/security/jwt/JwtUtils.java
@@ -8,7 +8,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
 
-import com.oopproject.form.services.UserDetailsImpl;
+import com.oopproject.form.service.UserDetailsImpl;
 
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.Jwts;
