+++ backend/src/main/java/com/oopproject/form/service/UserService.java
@@ -1,4 +1,4 @@
-package com.oopproject.form.services;
+package com.oopproject.form.service;
 
 import java.util.List;
